backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
supabase==2.28.3
python-dotenv==1.2.2
google-genai>=1.0.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
google-api-python-client==2.194.0
pdf2image==1.17.0
Pillow==12.2.0
apscheduler==3.11.2
python-multipart==0.0.26
httpx==0.28.1
pydantic-settings==2.14.0
email-validator==2.2.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
